LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/about-blank-popup.https-expected.txt


PASS Cross-Origin-Embedder-Policy is inherited by about:blank popup.
